Remove support for symlinks and empty directories
This commit changes the database schema to remove support for storing
symbolic links and empty directories in the history. Because the target
of a symbolic link can be an absolute path, symbolic links committed on
one platform cannot be checked out on another platform. While being able
to save empty directories would be a nice feature, it isn't important
enough to justify complicating the database schema.