Add timelines to database schema
Previously, the intention was for there to be a separate database for
each timeline. This commit changes the database schema so that
information from each timeline and information about the timelines
themselves is all stored in a single database.

Storing all the information in a single database allows for efficiently
determining which blobs are not being referenced by any files so that
they can be cleaned up. It also eliminates duplicate data which would
otherwise occur between databases.