Simplify database schema
Previously, the database schema tracked the parent-child relationships
between paths. The problem with this is that the program doesn't track
directories, so the paths of parent directories would have to be added
to the database when files are added and removed when those files are
removed. This would add unnecessary complexity to the database.

As of this commit, the database no longer tracks the relationships
between paths. This feature was only added so that it would be possible
to find all the descendants of a path, and this can still be done out
of the database or in the database using LIKE.