Fix handling of relative paths
Adding a file to a snapshot requires both its absolute path and its path
relative to its working directory. The absolute path is needed for
reading the file contents, and the relative path is needed so that the
file can be restored to another working directory.

Previously, methods for adding files to snapshots only accepted one
path.