Make repositories portable
This commit removes some features to make the program easier to use and
allow repositories to be synchronized between machines.

Repositories are now linked to a single working directory, and they are
stored inside their working directory. It is no longer possible for
multiple working directories to share a repository or for repositories
to be stored in a different location from their working directory.

This change greatly simplifies the interface and also makes it possible
for working directories—with their repositories—to be easily
synchronized between machines using third-party file synchronization
software. This new design is meant to be simpler and easier for a user
to understand.

Some of the features that were removed could still be added again in a
future version.