Remove the ability to import/export repositories
This commit removes the ability to import and export repositories as a
file. The original purpose of the feature was to allow for backing up
repositories. Now that repositories are stored in their working
directories, they can be backed up with the working directory itself
however the user chooses to do so.