Simplify input fields
This commit changes the way the data in the input form is saved. Now,
the information is only saved whenever one of the input fields is
activated. Previously, information was also saved when the focus
changed.

The purpose of this change is to simplify the UI controller and to make
the behavior of the program more predictable for the user.