Replace file globbing with literal path matching
Previously, it was possible to ignore files using shell globbing
patterns. Now, paths in the ignore file are interpreted literally so
that only specific files or directories can be ignored.

The reason for this change is that the behavior of globbing patterns is
inconsistent between platforms. Also, this change makes the system for
ignoring files simpler and easier for the user to understand. Most users
don't know how to use globbing patterns, and hiding this behavior in the
UI is difficult if there's no easy way to escape globbing pattern
automatically.