Make custom controls compatible with Scene Builder
This commit makes changes to the custom JavaFX controls used in the GUI
to make them compatible with Scene Builder. It also adds kotlin-reflect
as a dependency, which needs to be bundled with the program in a fat jar
in order for Scene Builder to be able to import it.