Prevent daemon from making unnecessary snapshots
Previously, the daemon would create a snapshot whenever the database
file was modified, resulting in hundreds of empty snapshots. This commit
adds the ability to limit which files the daemon watches for changes. It
still watches directories excluded by the user because re-reading the
exclude pattern file on each change would be expensive.