Allow for destructively opening a repository
This commit changes the API to allow for destructively opening a
repository that otherwise may not be able to be opened. The purpose of
this change is to give the program the chance to recover from
a corrupt repository while also making clear to the user that data may
be lost in the process. This can be used to implement restoring from a
backup of the metadata.