Remove system service support
This commit removes support for launching the daemon as a Windows
service, systemd service, or Launchd agent. The current plan is to
distribute the program as a self-contained fat jar, and there's no easy
way to have a service start the program on boot unless the
executable/jar is installed to a static path. This feature may be
re-added in the future, but for now, it is up to the user to put a
link/shortcut to the jar in their application startup directory.