Change startup behavior of UI and daemon
Previously, closing and re-opening the UI would hide the primary stage
and then show it again. Now, Closing and re-opening the UI closes and
re-creates the primary stage. This means that restarting the UI will
re-run much of the initialization code, which can be useful for dealing
with bugs.

This commit also ensures that the entire JVM will exit if JavaFX fails
to start. Previously, if JavaFX failed to start the first time the
program was run, subsequent invocations wouldn't attempt to start it.