Ensure database runs integrity check on connection
Previously, the database would only run an integrity check the first
time it was connected to per JVM. Now, the integrity check is run each
time the database is connected to, even if it's been connected to
before. This helps inform users of database corruption early so they can
repair it.